+++ examples/depth_filtered.rs
@@ -130,7 +130,7 @@ fn main() -> anyhow::Result<()> {
         &rerun::Pinhole::new(
             rerun::components::PinholeProjection::from_focal_length_and_principal_point(
                 [intrinsic.fx, intrinsic.fy],
-                [intrinsic.cx as f32, intrinsic.cy as f32],
+                [intrinsic.cx, intrinsic.cy],
             ),
         ),
     )?;
@@ -141,7 +141,7 @@ fn main() -> anyhow::Result<()> {
         rerun::Pinhole::new(
             rerun::components::PinholeProjection::from_focal_length_and_principal_point(
                 [intrinsic.fx / f, intrinsic.fy / f],
-                [intrinsic.cx as f32 / f, intrinsic.cy as f32 / f],
+                [intrinsic.cx / f, intrinsic.cy / f],
             ),
         )
     })?;
